xfs_fsr: correct type in fsrprintf() call
Use %lld instead of %d for howlong variable.

Coverity-id: 1596598

+++ fsr/xfs_fsr.c
@@ -426,7 +426,8 @@ fsrallfs(char *mtab, time_t howlong, char *leftofffile)
 	fsdesc_t *fsp;
 	struct stat sb, sb2;
 
-	fsrprintf("xfs_fsr -m %s -t %d -f %s ...\n", mtab, howlong, leftofffile);
+	fsrprintf("xfs_fsr -m %s -t %lld -f %s ...\n", mtab,
+		  (long long)howlong, leftofffile);
 
 	endtime = starttime + howlong;
 	fs = fsbase;
